3 more configurable ragnarok cards

+++ src/main/java/ragnarok_cards/Utils/VerificateCards.java
@@ -43,7 +43,7 @@ public class VerificateCards {
         cardNameToChance.put("witch",3);
     }
 
-    static public boolean passCheck(int quantity, float checkMultiplier){
+    static public boolean passCheck(int quantity, double checkMultiplier){
         return (rand.nextInt(100 ) < quantity * checkMultiplier );
     }
     static public boolean passCheck(String cardName,  Map<String, Integer> cards){
@@ -84,7 +84,7 @@ public class VerificateCards {
         return quantity;
     }
 
-    static public boolean SingleCardActivate (PlayerEntity player, String cardName, float multiplier){
+    static public boolean SingleCardActivate (PlayerEntity player, String cardName, double multiplier){
         int quantity = HowManyCards(player, cardName);
         return passCheck(quantity,cardNameToChance.get(cardName) * multiplier);
     }
